Rename VisibleBehaviour to Vision component, namespace and move it
Also move VoiceComponent.
