+++ src/FieldWarning/Assets/Units/Component/Sound.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6be14dabb55a32a409bd635676e40790
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
