+++ src/FieldWarning/Assets/Units/Component/Sound/VoiceComponent.cs
