+++ src/FieldWarning/Assets/Units/Component/Vision.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fc0f5de446d27ab4db0cb3f64689e891
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
